+++ ex4nicegui/reactive/officials/tab_panels.py
@@ -1,4 +1,3 @@
-import inspect
 from typing import Any, Awaitable, Callable, Optional, Union
 from weakref import WeakValueDictionary
 from ex4nicegui.reactive.services.reactive_service import ParameterClassifier
@@ -8,9 +7,10 @@ from ex4nicegui.utils.signals import (
     _TMaybeRef as TMaybeRef,
 )
 from ex4nicegui.utils.scheduler import next_tick
-from nicegui import ui, background_tasks, core
+from nicegui import ui
 from .base import BindableUi
 from ex4nicegui.reactive.mixins.value_element import ValueElementMixin
+from .tab_panel import lazy_tab_panel
 
 
 class TabPanelsBindableUi(BindableUi[ui.tab_panels], ValueElementMixin[bool]):
@@ -57,21 +57,6 @@ class TabPanelsBindableUi(BindableUi[ui.tab_panels], ValueElementMixin[bool]):
         return super().bind_prop(prop, value)
 
 
-class lazy_tab_panel(ui.tab_panel):
-    def __init__(self, name: str) -> None:
-        super().__init__(name)
-        self._build_fn = None
-
-    def try_run_build_fn(self):
-        if self._build_fn:
-            _helper.run_build_fn(self, self._props["name"])
-            self._build_fn = None
-
-    def build_fn(self, fn: Callable[..., Union[None, Awaitable]]):
-        self._build_fn = fn
-        return fn
-
-
 class LazyTabPanelsBindableUi(TabPanelsBindableUi):
     def __init__(
         self,
@@ -96,65 +81,40 @@ class LazyTabPanelsBindableUi(TabPanelsBindableUi):
             value, on_change=on_change, animated=animated, keep_alive=keep_alive
         )
 
-        self.__panels: WeakValueDictionary[str, lazy_tab_panel] = WeakValueDictionary()
+        self._panels: WeakValueDictionary[str, lazy_tab_panel] = WeakValueDictionary()
 
         if value:
 
             @self._ui_effect
             def _():
                 current_value = to_value(value)
-                if current_value in self.__panels:
-                    panel = self.__panels[current_value]
+                if current_value in self._panels:
+                    panel = self._panels[current_value]
 
                     @next_tick
                     def _():
                         panel.try_run_build_fn()
 
     def add_tab_panel(self, name: str):
-        def decorator(fn: Callable[..., Union[None, Awaitable]]):
-            with self:
-                panel = lazy_tab_panel(name)
-            str_name = panel._props["name"]
-            self.__panels[str_name] = panel
-            panel.build_fn(fn)
-
-            if self.value == name:
-                panel.try_run_build_fn()
-
-            return panel
-
-        return decorator
+        return TabPanelDescriptor(self, name)
 
 
-class _helper:
-    @staticmethod
-    def run_build_fn(panel: lazy_tab_panel, name: str) -> None:
-        """ """
-        fn = panel._build_fn
-        if fn is None:
-            return
-        try:
-            expects_arguments = any(
-                p.default is inspect.Parameter.empty
-                and p.kind is not inspect.Parameter.VAR_POSITIONAL
-                and p.kind is not inspect.Parameter.VAR_KEYWORD
-                for p in inspect.signature(fn).parameters.values()
-            )
+class TabPanelDescriptor:
+    def __init__(self, tab_panels: LazyTabPanelsBindableUi, panle_name: str):
+        self.panle_name = panle_name
+        self.tab_panels = tab_panels
 
-            with panel:
-                result = fn(name) if expects_arguments else fn()
-            if isinstance(result, Awaitable):
-                # NOTE: await an awaitable result even if the handler is not a coroutine (like a lambda statement)
-                async def wait_for_result():
-                    with panel:
-                        try:
-                            await result
-                        except Exception as e:
-                            core.app.handle_exception(e)
+    def __call__(self, fn: Callable[..., Union[None, Awaitable]]):
+        with self.tab_panels:
+            panel = lazy_tab_panel(self.panle_name)
+        str_name = panel.element._props["name"]
+        self.tab_panels._panels[str_name] = panel
+        panel.build_fn(fn)
 
-                if core.loop and core.loop.is_running():
-                    background_tasks.create(wait_for_result(), name=str(fn))
-                else:
-                    core.app.on_startup(wait_for_result())
-        except Exception as e:
-            core.app.handle_exception(e)
+        if self.tab_panels.value == self.panle_name:
+            panel.try_run_build_fn()
+
+        return panel
+
+    def props(self, **kwargs):
+        pass
